pyproject.toml
[project]
name = "dac-hrag"
version = "0.1.0"
description = "Distributed Attributed Community-Hierarchical RAG system"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "numpy>=1.26.0",
    "faiss-cpu>=1.7.4",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.4",
    "asyncpg>=0.29.0",
    "igraph>=0.11.3",
    "leidenalg>=0.10.2",
    "scikit-learn>=1.4.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
    "tiktoken>=0.5.0",
    "pyyaml>=6.0.1",
    "msgpack>=1.0.7",
    "sentence-transformers>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "types-protobuf>=4.24.0",
    "types-pyyaml>=6.0.12",
]

[project.scripts]
dachrag = "src.api.cli:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*", "scripts*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow",
    "integration: marks integration tests",
    "distributed: marks distributed tests",
]
pythonpath = ["."]
